fix(dbengine): prevent negative tier retention reporting (#23648)
* fix(api): update retention condition checks to prevent incorrect time publishing

* fix(rrd-retention): increase maximum expected retention to accommodate realistic long-term estimates

* fix(journalfile, rrd-retention): improve comments for clarity and adjust retention logic

(cherry picked from commit c0de72d2abb09d55eaeaa8a1341408f5ed14d4cd)

+++ src/libnetdata/inicfg/inicfg_api.c
@@ -415,7 +415,7 @@ static STRING *reformat_duration_days_to_seconds(STRING *value) {
 
 time_t inicfg_get_duration_days_to_seconds(struct config *root, const char *section, const char *name, unsigned default_value_seconds) {
     char default_str[128];
-    duration_snprintf(default_str, sizeof(default_str), (int)default_value_seconds, "s", false);
+    duration_snprintf(default_str, sizeof(default_str), (int64_t)default_value_seconds, "s", false);
 
     struct config_option *opt = inicfg_get_raw_value(
         root, section, name, default_str,
